requirements_working_app.txt
PyMuPDF==1.23.8
openai==1.76.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.12.0
dnspython==2.7.0
PyPDF2==3.0.1
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
anthropic>=0.18.1
httpx==0.28.1
